refactor: move Dialog to global components directory
